feat(combobox): deplacer dans le parent le setVisibleOptions update
