+++ src/client/components/ui/Form/Combobox/ComboboxContext.tsx
@@ -1,4 +1,4 @@
-import React, { Dispatch, ReducerState, SetStateAction,useContext } from 'react';
+import React, { Dispatch, ReducerState,useContext } from 'react';
 
 import NoProviderError from '~/client/Errors/NoProviderError';
 
@@ -9,7 +9,7 @@ type ContextContent = {
 	dispatch: Dispatch<ComboboxAction>,
 	onOptionSelection: (option: Element) => void,
 	filter: (element: Element, value: string) => boolean,
-	setVisibleOptions: Dispatch<SetStateAction<Array<string>>>
+	onUpdateVisibleOptions: (option: Element) => void,
 	visibleOptions: Array<string>
 }
